Make beauty

+++ TaDS2/sort.h
@@ -7,7 +7,7 @@ unsigned long long tick(void);
 void measure_sorting_time();
 
 void swap_students(struct Student *a, struct Student *b);
-void mysort_stud(struct Student *base, size_t nitems);
+void mysort_stud(struct Student *base, int nitems);
 void swap_keys(struct Key *a, struct Key *b);
 void mysort_keys(struct Key *base, size_t nitems);
